requirements.txt
numpy
faiss-cpu
pymupdf
pytesseract
python-dotenv
requests
streamlit
tqdm
sentence-transformers
transformers
langchain
pypdf
pdfplumber
chromadb
langchain-groq
langchain-community
langchain-huggingface
langchain-chroma